requirements/test.txt
# Test dependencies go here.
-r base.txt

factory-boy==2.8.1
coverage==4.3.1
django_coverage_plugin==1.3.1


